+++ td/telegram/files/FileHashUploader.cpp
@@ -66,12 +66,11 @@ Status FileHashUploader::loop_impl() {
     TRY_STATUS(loop_sha());
   }
   if (state_ == State::NetRequest) {
-    // messages.getDocumentByHash#338e2464 sha256:bytes size:int mime_type:string = Document;
+    // messages.getDocumentByHash#338e2464 sha256:bytes size:long mime_type:string = Document;
     auto hash = BufferSlice(32);
     sha256_state_.extract(hash.as_slice(), true);
     auto mime_type = MimeType::from_extension(PathView(local_.path_).extension(), "image/gif");
-    auto query =
-        telegram_api::messages_getDocumentByHash(std::move(hash), static_cast<int32>(size_), std::move(mime_type));
+    auto query = telegram_api::messages_getDocumentByHash(std::move(hash), size_, std::move(mime_type));
     LOG(INFO) << "Send getDocumentByHash request: " << to_string(query);
     auto ptr = G()->net_query_creator().create(query);
     G()->net_query_dispatcher().dispatch_with_callback(std::move(ptr), actor_shared(this));
